cambios post en test e2e de put products, put y delete de products

+++ controller/products.js
@@ -5,13 +5,7 @@ const db = connect();
 module.exports = {
  postProducts: async (req, resp, next) => {
     const { name, price, image, type } = req.body;
-    const newProduct = {
-      name,
-      price,
-      image,
-      type,
-    };
-
+    
     try {
       
       const products = db.collection('product');
@@ -24,6 +18,13 @@ module.exports = {
       if (!name || !price) {
         return resp.status(400).json({ error: 'tienes que ingresar todos los datos' });
       }
+      const newProduct = {
+        name,
+        price,
+        image,
+        type,
+      };
+  
       
       await products.insertOne(newProduct);
      
@@ -85,32 +86,45 @@ module.exports = {
     try {
         
       const products = db.collection('product');
-
       const  productsId = req.params.productId;
+      console.log({products});
+      console.log({productsId});
       
       if (!/^[0-9a-fA-F]{24}$/.test(productsId)) {
         return resp.status(404).json({ error: 'El ID del producto solicitado no es válido' });
       };
 
       let query = { _id: new ObjectId(productsId) };
+      console.log({query});
 
       const productData = await products.findOne(query);
+      console.log({productData});
 
       if (!productData) {
         return resp.status(404).json({ error: 'El producto solicitado no existe' });
       }
       
       const body = req.body;
+      console.log({body});
 
       if (!body || Object.keys(body).length === 0) {
         return resp.status(400).json({ error: 'Debe haber al menos una propiedad para actualizar' });
       }
 
+      
+      if (body.hasOwnProperty('price')) {
+        const price = body.price;
+        if (typeof price !== 'number') {// Verificar si el precio no es un número
+          return resp.status(400).json({ error: 'El precio debe ser un número' });
+        }
+      }
+    
+
       const productUpdate = await products.updateOne(query, { $set: body });
-
-      resp.json({ productUpdate, message: 'El producto ha sido actualizado' });
-
-
+      console.log({productUpdate});
+      const updatedProduct = await products.findOne(query);
+      resp.status(200).json(updatedProduct);
+          
     } catch (error) {
         console.error(error);
         resp.status(500).json({ error: 'Error al actualizar el producto' });
@@ -120,16 +134,24 @@ module.exports = {
   deleteProducts: async ( req, resp, next) => {
     try{
         
-        const products = db.collection('product');
-        const  productsId = req.params.productId;
-  
-        let query = { _id: new ObjectId(productsId) };
-  
-        const productData = await products.findOne(query);
-  
-        if (!productData) {
-          return resp.status(404).json({ error: 'El producto solicitado no existe' });
-        }
+      const products = db.collection('product');
+      const  productsId = req.params.productId;
+      console.log({products});
+      console.log({productsId});
+      
+      if (!/^[0-9a-fA-F]{24}$/.test(productsId)) {
+        return resp.status(404).json({ error: 'El ID del producto solicitado no es válido' });
+      };
+
+      let query = { _id: new ObjectId(productsId) };
+      console.log({query});
+
+      const productData = await products.findOne(query);
+      console.log({productData});
+
+      if (!productData) {
+        return resp.status(404).json({ error: 'El producto solicitado no existe' });
+      }
 
         const productDelete = await products.deleteOne(query);
